complete code till chating Engine

+++ controllers/api/v1/posts_api.js
@@ -26,17 +26,18 @@ module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
         //.id means converting the ._id in string
-        
+        if(post.user == req.user.id){
             post.remove();
             await Comment.deleteMany({ post: req.params.id });
             return res.status(200).json({
                 message:"post delete successfully!!!"
             })
         
-        // } else {
-        //     req.flash('error', 'you cannot delete this Post!');
-        //     return res.redirect('back');
-        // }
+        } else {
+            return res.status(401).json({
+                message:"you cannot delete this post"
+            })
+        }
     } catch (err) {
         return res.json(500,{
             message:"Internal Server Error!"
